feat: Extract basic date formatting building blocs to their own util file and package
