+++ app/src/main/java/com/infomaniak/mail/utils/MailDateFormatUtils.kt
@@ -19,9 +19,11 @@ package com.infomaniak.mail.utils
 
 import android.content.Context
 import android.os.Build
-import android.text.format.DateFormat
 import com.infomaniak.lib.core.utils.*
 import com.infomaniak.mail.R
+import com.infomaniak.mail.utils.date.DateFormatUtils.formatTime
+import com.infomaniak.mail.utils.date.DateFormatUtils.fullDateWithYear
+import com.infomaniak.mail.utils.date.DateFormatUtils.fullDateWithoutYear
 import java.time.format.FormatStyle
 import java.util.Date
 
@@ -29,11 +31,11 @@ object MailDateFormatUtils {
 
     fun Context.mailFormattedDate(date: Date): CharSequence = with(date) {
         return when {
-            isToday() -> format(localHourFormat())
+            isToday() -> formatTime(this)
             isYesterday() -> getString(
                 R.string.messageDetailsDateAt,
                 getString(R.string.messageDetailsYesterday),
-                format(localHourFormat()),
+                formatTime(this),
             )
             isThisYear() -> fullDateWithYear(this)
             else -> fullDateWithoutYear(date = this)
@@ -47,32 +49,4 @@ object MailDateFormatUtils {
             format(FORMAT_DATE_DAY_FULL_MONTH_YEAR_WITH_TIME)
         }
     }
-}
-
-// Do not use the 12/24 hours format directly. Call localHourFormat() instead
-private const val FORMAT_DATE_24HOUR = "HH:mm"
-private const val FORMAT_DATE_12HOUR = "hh:mm a"
-private const val FORMAT_DATE_WITH_YEAR = "d MMM yyyy"
-private const val FORMAT_DATE_WITHOUT_YEAR = "d MMM"
-
-fun Context.fullDateWithYear(date: Date): String {
-    return date.formatDateTime(this, FORMAT_DATE_WITHOUT_YEAR, localHourFormat())
-}
-
-fun Context.fullDateWithoutYear(date: Date): String {
-    return date.formatDateTime(this, FORMAT_DATE_WITH_YEAR, localHourFormat())
-}
-
-fun Context.dayOfWeekDate(date: Date): String {
-    return date.formatDateTime(this, FORMAT_DATE_DAY_MONTH, localHourFormat())
-}
-
-private fun Date.formatDateTime(context: Context, dateFormat: String, timeFormat: String) = context.getString(
-    R.string.messageDetailsDateAt,
-    format(dateFormat),
-    format(timeFormat),
-)
-
-private fun Context.localHourFormat(): String {
-    return if (DateFormat.is24HourFormat(this)) FORMAT_DATE_24HOUR else FORMAT_DATE_12HOUR
 }
